requirements.txt
# UMSF-Net Dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Data processing
rasterio>=1.3.0
pillow>=9.4.0

# Machine learning
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Logging and monitoring
tensorboard>=2.12.0
tqdm>=4.65.0

# Utilities
einops>=0.6.0

# Optional: for t-SNE visualization
# scikit-learn already includes TSNE

# Optional: for advanced augmentation
# albumentations>=1.3.0
